backwards-compatible change to support not using dumpcap
At least with the Windows implementation of tshark (specifically, TShark (Wireshark) 2.2.2 (v2.2.2-0-g775fb08)), the combination of reading from stdin ("-r -") and writing to an output file that is a ring buffer ("-b" options) produces the error below.

  tshark: Multiple capture files requested, but a capture isn't being done.

The only way I found to workaround this was to not read from stdin.  To do so, I had to remove the use of dumpcap within the live_capture class, instead opting to start the tshark process with the parameters set such that it performs the capture itself.

This is controlled via a new constructor argument, use_dumpcap, which defaults to True for backwards compatibility.